+++ internal/tui/bounds_test.go
@@ -34,11 +34,11 @@ func TestBoundsCentroid(t *testing.T) {
 func TestBoundsExtend(t *testing.T) {
 	a := tui.NewBounds(3, 4, 5, 6)
 	b := tui.NewBounds(1, 2, 12, 13)
-	a.Extend(b)
+	a.Extend(&b)
 	assert.Equal(t, b, a)
 
 	a = tui.NewBounds(3, 4, 5, 6)
 	b = tui.NewBounds(5, 6, 12, 13)
-	a.Extend(b)
+	a.Extend(&b)
 	assert.Equal(t, tui.NewBounds(3, 4, 14, 15), a)
 }