+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [[unpublished]](https://github.com/mlange-42/bbn/compare/v0.2.0...main)
+
+### Features
+
+* Add support for selecting outcomes by mouse click (#33)
+
 ## [[v0.2.0]](https://github.com/mlange-42/bbn/compare/v0.1.0...v0.2.0)
 
 ### Features